chore(engine): migrate io.zeebe.broker.engine tests
