+++ engine/src/test/java/io/zeebe/engine/processor/workflow/instance/WorkflowInstanceFunctionalTest.java
@@ -1,5 +1,5 @@
 /*
- * Zeebe Broker Core
+ * Zeebe Workflow Engine
  * Copyright © 2017 camunda services GmbH (info@camunda.com)
  *
  * This program is free software: you can redistribute it and/or modify
@@ -15,12 +15,12 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package io.zeebe.broker.engine;
+package io.zeebe.engine.processor.workflow.instance;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.zeebe.broker.test.EmbeddedBrokerRule;
+import io.zeebe.engine.util.EngineRule;
 import io.zeebe.exporter.api.record.Assertions;
 import io.zeebe.exporter.api.record.Record;
 import io.zeebe.exporter.api.record.RecordMetadata;
@@ -33,65 +33,40 @@ import io.zeebe.model.bpmn.Bpmn;
 import io.zeebe.model.bpmn.BpmnModelInstance;
 import io.zeebe.protocol.BpmnElementType;
 import io.zeebe.protocol.ExecuteCommandResponseDecoder;
-import io.zeebe.protocol.ValueType;
-import io.zeebe.protocol.impl.record.value.workflowinstance.WorkflowInstanceCreationRecord;
-import io.zeebe.protocol.intent.DeploymentIntent;
 import io.zeebe.protocol.intent.JobIntent;
 import io.zeebe.protocol.intent.WorkflowInstanceCreationIntent;
 import io.zeebe.protocol.intent.WorkflowInstanceIntent;
-import io.zeebe.test.broker.protocol.commandapi.CommandApiRule;
-import io.zeebe.test.broker.protocol.commandapi.ExecuteCommandResponse;
-import io.zeebe.test.broker.protocol.commandapi.PartitionTestClient;
-import io.zeebe.test.util.MsgPackUtil;
 import io.zeebe.test.util.Strings;
 import io.zeebe.test.util.record.RecordingExporter;
-import io.zeebe.test.util.record.RecordingExporterTestWatcher;
 import java.io.File;
 import java.net.URISyntaxException;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import org.agrona.DirectBuffer;
 import org.assertj.core.util.Files;
-import org.junit.BeforeClass;
 import org.junit.ClassRule;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.RuleChain;
 
 public class WorkflowInstanceFunctionalTest {
 
-  private static final EmbeddedBrokerRule BROKER_RULE = new EmbeddedBrokerRule();
-  private static final CommandApiRule API_RULE = new CommandApiRule(BROKER_RULE::getAtomix);
-  private static PartitionTestClient testClient;
-
-  @ClassRule public static RuleChain ruleChain = RuleChain.outerRule(BROKER_RULE).around(API_RULE);
-
-  @Rule
-  public RecordingExporterTestWatcher recordingExporterTestWatcher =
-      new RecordingExporterTestWatcher();
-
-  @BeforeClass
-  public static void init() {
-    testClient = API_RULE.partitionClient();
-  }
+  @ClassRule public static final EngineRule ENGINE = new EngineRule();
 
   @Test
   public void shouldCreateWorkflowInstance() {
     // given
     final String processId = Strings.newRandomValidBpmnId();
     final long workflowKey =
-        testClient
-            .deployWorkflow(Bpmn.createExecutableProcess(processId).startEvent().endEvent().done())
-            .getKey();
-    final DirectBuffer variables = MsgPackUtil.asMsgPack("foo", "bar");
+        ENGINE
+            .deploy(Bpmn.createExecutableProcess(processId).startEvent().endEvent().done())
+            .getValue()
+            .getDeployedWorkflows()
+            .get(0)
+            .getWorkflowKey();
 
     // when
-    final WorkflowInstanceCreationRecord workflowInstance =
-        testClient.createWorkflowInstance(r -> r.setKey(workflowKey).setVariables(variables));
-    final long workflowInstanceKey = workflowInstance.getInstanceKey();
+    final long workflowInstanceKey =
+        ENGINE.workflowInstance().ofBpmnProcessId(processId).withVariable("foo", "bar").create();
 
     // then
     final long workflowCompletedPosition =
@@ -140,13 +115,10 @@ public class WorkflowInstanceFunctionalTest {
     // given
     final String processId = Strings.newRandomValidBpmnId();
     final String startId = Strings.newRandomValidBpmnId();
-    testClient.deploy(
-        Bpmn.createExecutableProcess(processId).startEvent(startId).endEvent().done());
+    ENGINE.deploy(Bpmn.createExecutableProcess(processId).startEvent(startId).endEvent().done());
 
     // when
-    final WorkflowInstanceCreationRecord workflowInstance =
-        testClient.createWorkflowInstance(r -> r.setBpmnProcessId(processId));
-    final long workflowInstanceKey = workflowInstance.getInstanceKey();
+    final long workflowInstanceKey = ENGINE.workflowInstance().ofBpmnProcessId(processId).create();
 
     // then
     final Record<WorkflowInstanceCreationRecordValue> workflowCreated =
@@ -176,7 +148,7 @@ public class WorkflowInstanceFunctionalTest {
     // given
     final String processId = Strings.newRandomValidBpmnId();
     final String sequenceId = Strings.newRandomValidBpmnId();
-    testClient.deploy(
+    ENGINE.deploy(
         Bpmn.createExecutableProcess(processId)
             .startEvent()
             .sequenceFlowId(sequenceId)
@@ -184,8 +156,7 @@ public class WorkflowInstanceFunctionalTest {
             .done());
 
     // when
-    final long workflowInstanceKey =
-        testClient.createWorkflowInstance(r -> r.setBpmnProcessId(processId)).getInstanceKey();
+    final long workflowInstanceKey = ENGINE.workflowInstance().ofBpmnProcessId(processId).create();
 
     // then
     final Record<WorkflowInstanceRecordValue> sequenceFlow =
@@ -209,11 +180,10 @@ public class WorkflowInstanceFunctionalTest {
     // given
     final String processId = Strings.newRandomValidBpmnId();
     final String endId = Strings.newRandomValidBpmnId();
-    testClient.deploy(Bpmn.createExecutableProcess(processId).startEvent().endEvent(endId).done());
+    ENGINE.deploy(Bpmn.createExecutableProcess(processId).startEvent().endEvent(endId).done());
 
     // when
-    final long workflowInstanceKey =
-        testClient.createWorkflowInstance(r -> r.setBpmnProcessId(processId)).getInstanceKey();
+    final long workflowInstanceKey = ENGINE.workflowInstance().ofBpmnProcessId(processId).create();
 
     // then
     final Record<WorkflowInstanceRecordValue> endEvent =
@@ -244,11 +214,10 @@ public class WorkflowInstanceFunctionalTest {
             .endEvent()
             .done();
 
-    testClient.deploy(model);
+    ENGINE.deploy(model);
 
     // when
-    final long workflowInstanceKey =
-        testClient.createWorkflowInstance(r -> r.setBpmnProcessId(processId)).getInstanceKey();
+    final long workflowInstanceKey = ENGINE.workflowInstance().ofBpmnProcessId(processId).create();
 
     // then
     final Record<WorkflowInstanceRecordValue> activityReady =
@@ -282,7 +251,7 @@ public class WorkflowInstanceFunctionalTest {
     final String processId = Strings.newRandomValidBpmnId();
     final String taskId = Strings.newRandomValidBpmnId();
     final String taskType = Strings.newRandomValidBpmnId();
-    testClient.deploy(
+    ENGINE.deploy(
         Bpmn.createExecutableProcess(processId)
             .startEvent()
             .serviceTask(taskId, t -> t.zeebeTaskType(taskType).zeebeTaskRetries(5))
@@ -290,8 +259,7 @@ public class WorkflowInstanceFunctionalTest {
             .done());
 
     // when
-    final long workflowInstanceKey =
-        testClient.createWorkflowInstance(r -> r.setBpmnProcessId(processId)).getInstanceKey();
+    final long workflowInstanceKey = ENGINE.workflowInstance().ofBpmnProcessId(processId).create();
 
     // then
     final Record<WorkflowInstanceRecordValue> activityActivated =
@@ -323,7 +291,7 @@ public class WorkflowInstanceFunctionalTest {
     final String processId = Strings.newRandomValidBpmnId();
     final String taskId = Strings.newRandomValidBpmnId();
     final String taskType = Strings.newRandomValidBpmnId();
-    testClient.deploy(
+    ENGINE.deploy(
         Bpmn.createExecutableProcess(processId)
             .startEvent()
             .serviceTask(
@@ -333,8 +301,7 @@ public class WorkflowInstanceFunctionalTest {
             .done());
 
     // when
-    final long workflowInstanceKey =
-        testClient.createWorkflowInstance(r -> r.setBpmnProcessId(processId)).getInstanceKey();
+    final long workflowInstanceKey = ENGINE.workflowInstance().ofBpmnProcessId(processId).create();
 
     // then
     final Record<JobRecordValue> event =
@@ -366,14 +333,12 @@ public class WorkflowInstanceFunctionalTest {
             .endEvent()
             .done();
 
-    testClient.deploy(definition);
+    ENGINE.deploy(definition);
 
-    final long workflowInstanceKey =
-        testClient.createWorkflowInstance(r -> r.setBpmnProcessId(processId)).getInstanceKey();
+    final long workflowInstanceKey = ENGINE.workflowInstance().ofBpmnProcessId(processId).create();
 
     // when
-    testClient.activateAndCompleteFirstJob(
-        taskType, r -> r.getJobHeaders().getWorkflowInstanceKey() == workflowInstanceKey);
+    ENGINE.job().ofInstance(workflowInstanceKey).withType(taskType).complete();
 
     // then
     final Record<WorkflowInstanceRecordValue> activityActivatedEvent =
@@ -384,7 +349,7 @@ public class WorkflowInstanceFunctionalTest {
             .withElementId(taskId)
             .getFirst();
     final Record<JobRecordValue> jobCompleted =
-        testClient.receiveFirstJobEvent(JobIntent.COMPLETED);
+        RecordingExporter.jobRecords().withIntent(JobIntent.COMPLETED).getFirst();
     final Record<WorkflowInstanceRecordValue> activityCompleting =
         RecordingExporter.workflowInstanceRecords()
             .withWorkflowInstanceKey(workflowInstanceKey)
@@ -411,7 +376,6 @@ public class WorkflowInstanceFunctionalTest {
         .hasWorkflowInstanceKey(workflowInstanceKey);
   }
 
-  // todo(npepinpe): is this a useful test?
   @Test
   public void testWorkflowInstanceStatesWithServiceTask() {
     // given
@@ -427,14 +391,11 @@ public class WorkflowInstanceFunctionalTest {
             .endEvent(endId)
             .done();
 
-    testClient.deploy(definition);
-
-    final long workflowInstanceKey =
-        testClient.createWorkflowInstance(r -> r.setBpmnProcessId(processId)).getInstanceKey();
+    ENGINE.deploy(definition);
+    final long workflowInstanceKey = ENGINE.workflowInstance().ofBpmnProcessId(processId).create();
 
     // when
-    testClient.activateAndCompleteFirstJob(
-        taskType, r -> r.getJobHeaders().getWorkflowInstanceKey() == workflowInstanceKey);
+    ENGINE.job().ofInstance(workflowInstanceKey).withType(taskType).complete();
 
     // then
     final List<Record<WorkflowInstanceRecordValue>> workflowEvents =
@@ -480,25 +441,13 @@ public class WorkflowInstanceFunctionalTest {
     deploymentResource.put("resourceType", ResourceType.YAML_WORKFLOW);
     deploymentResource.put("resourceName", "simple-workflow.yaml");
 
-    final ExecuteCommandResponse deploymentResp =
-        API_RULE
-            .createCmdRequest()
-            .type(ValueType.DEPLOYMENT, DeploymentIntent.CREATE)
-            .command()
-            .put("resources", Collections.singletonList(deploymentResource))
-            .done()
-            .sendAndAwait();
+    ENGINE.deployment().withYamlResource(yamlWorkflow.getBytes(UTF_8)).deploy();
 
-    testClient.receiveFirstDeploymentEvent(DeploymentIntent.CREATED, deploymentResp.getKey());
-
-    final long workflowInstanceKey =
-        testClient.createWorkflowInstance(r -> r.setBpmnProcessId(processId)).getInstanceKey();
+    final long workflowInstanceKey = ENGINE.workflowInstance().ofBpmnProcessId(processId).create();
 
     // when
-    testClient.activateAndCompleteFirstJob(
-        "foo", r -> r.getJobHeaders().getWorkflowInstanceKey() == workflowInstanceKey);
-    testClient.activateAndCompleteFirstJob(
-        "bar", r -> r.getJobHeaders().getWorkflowInstanceKey() == workflowInstanceKey);
+    ENGINE.job().withType("foo").ofInstance(workflowInstanceKey).complete();
+    ENGINE.job().withType("bar").ofInstance(workflowInstanceKey).complete();
 
     // then
     final Record<WorkflowInstanceRecordValue> event =
