+++ engine/src/test/resources/workflows/collaboration.bpmn
