+++ engine/src/test/resources/workflows/invalid_process_condition.bpmn
